Reduce pixel count type size. Use static casts to convert to size_t

+++ firmware/components/hwSubsystem_thermalFrameGrabber/include/mlx90641.hpp
@@ -37,7 +37,7 @@ namespace externalHardwareSubsystem
         static constexpr float defaultBaselineTemperature{23.15};
         static constexpr float defaultEmissivity{0.95};
 
-        static constexpr size_t pixelCount{192};
+        static constexpr uint8_t pixelCount{192};
 
         /*Compile time assertions based on nature of the hardware*/
         static_assert(factorySetAddress==0x33U, "MLX90641 I2C address not default. Check datasheet");
